+++ mesh_to_sdf_client/src/gltf/scene/mod.rs
@@ -6,10 +6,9 @@ mod light;
 pub mod model;
 
 pub use camera::Camera;
+use gltf::scene::Node;
 pub use light::Light;
 pub use model::{GltfModel, Material};
-
-use gltf::scene::Node;
 
 use super::utils::transform_to_matrix;