+++ mesh_to_sdf/CHANGELOG.md
@@ -1,11 +1,11 @@
 # Changelog
 
-## Unreleased
+## [0.4.0] - 2024-09-17
 
 ### Added
 
 - New `AccelerationMethod`s with `Rtree` and `RtreeBvh`. `Rtree` uses a r-tree for distances and the normal sign method, while `RtreeBvh` uses a r-tree for distances and a bvh for raycast.
-    Both are about 4x faster than the previous `Bvh` for 10k and more queries. `RtreeBvh` is the new default method.
+    Both are about 4x faster than the previous `Bvh` for 10k and more queries.
 
 ### Changed
 
@@ -13,8 +13,10 @@
 - `AccelerationMethod` now includes the `SignMethod` when it makes sense. `generate_sdf` only takes the `AccelerationMethod` parameter, and the `SignMethod` is inferred from it.
     This is because not all acceleration methods support all sign methods. For example, `Rtree` only supports the normal sign method, while `RtreeBvh` supports raycasting only.
 - `Point` trait now requires `Debug` and `PartialEq` to be implemented.
+- `RtreeBvh` is the new default `AccelerationMethod`.
 
-## [0.3.0]
+
+## [0.3.0] - 2024-09-06
 
 ### Added